Require subscription_key and event_store
The subscritption key could be defined in two different ways
or not at all. This led to inconsistent code and non uniform names
of events in the database. By enforcing the subscription key to be
given when using the macro the developer can ensure that all
subcription names are consistent.

As this is a breaking change the version of the library is bumped from
2.x to 3.0